4.0.0 Pull Request

+++ extras-topo/src/main/java/com/metsci/glimpse/topo/io/TopoDataFile.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, Metron, Inc.
+ * Copyright (c) 2020, Metron, Inc.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -80,7 +80,7 @@ public class TopoDataFile
 
     public void requireValid( ) throws IOException
     {
-        long expectedSize = this.numRows * this.numCols * this.dataType.numBytes;
+        long expectedSize = ( long ) this.numRows * ( long ) this.numCols * this.dataType.numBytes;
         long actualSize = this.dataFile.length( );
 
         if ( actualSize == 0 && !this.dataFile.exists( ) )
